requirements.txt
git+https://github.com/revoltchat/revolt.py
asyncpg
git+https://github.com/bryanforbes/asyncpg-stubs
toml
lru-dict
